requirements.txt
fastapi>=0.134.0,<1.0
uvicorn[standard]>=0.41.0,<1.0
pydantic-settings>=2.13.0,<3.0
python-dotenv>=1.2.0,<2.0
sqlalchemy[asyncio]>=2.0.25,<3.0
aiosqlite>=0.19.0,<1.0
pytest>=9.0.0,<10.0
httpx>=0.28.0,<1.0
anyio>=4.0.0,<5.0
